Adaptando proyecto a patrón MVVM
