+++ Lector Excel/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:Lector_Excel"
-             StartupUri="MainWindow.xaml">
+             StartupUri="Views\MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary Source="pack://application:,,,/Fluent;Component/Themes/Generic.xaml"/>
     </Application.Resources>
