+++ Lector Excel/Views/Type1Window.xaml
@@ -14,7 +14,7 @@
                     <MenuItem.Icon>
                         <Rectangle>
                             <Rectangle.Fill>
-                                <ImageBrush ImageSource="media\common_icons.png" Viewbox="0,0,0.722,1" Stretch="None"/>
+                                <ImageBrush ImageSource="..\media\common_icons.png" Viewbox="0,0,0.722,1" Stretch="None"/>
                             </Rectangle.Fill>
                         </Rectangle>
                     </MenuItem.Icon>
@@ -23,7 +23,7 @@
                     <MenuItem.Icon>
                         <Rectangle>
                             <Rectangle.Fill>
-                                <ImageBrush ImageSource="media\common_icons.png" Viewbox="0,0,1.426,1" Stretch="None"/>
+                                <ImageBrush ImageSource="..\media\common_icons.png" Viewbox="0,0,1.426,1" Stretch="None"/>
                             </Rectangle.Fill>
                         </Rectangle>
 
@@ -37,7 +37,7 @@
                     <MenuItem.Icon>
                         <Rectangle>
                             <Rectangle.Fill>
-                                <ImageBrush ImageSource="media\common_icons.png" Viewbox="0,0,0.13,1" Stretch="None"/>
+                                <ImageBrush ImageSource="..\media\common_icons.png" Viewbox="0,0,0.13,1" Stretch="None"/>
                             </Rectangle.Fill>
                         </Rectangle>
                     </MenuItem.Icon>
@@ -46,7 +46,7 @@
                     <MenuItem.Icon>
                         <Rectangle>
                             <Rectangle.Fill>
-                                <ImageBrush ImageSource="media\common_icons.png" Viewbox="0,0,0.166,1" Stretch="None"/>
+                                <ImageBrush ImageSource="..\media\common_icons.png" Viewbox="0,0,0.166,1" Stretch="None"/>
                             </Rectangle.Fill>
                         </Rectangle>
                     </MenuItem.Icon>
@@ -55,7 +55,7 @@
                     <MenuItem.Icon>
                         <Rectangle>
                             <Rectangle.Fill>
-                                <ImageBrush ImageSource="media\common_icons.png" Viewbox="0,0,0.907,1" Stretch="None"/>
+                                <ImageBrush ImageSource="..\media\common_icons.png" Viewbox="0,0,0.907,1" Stretch="None"/>
                             </Rectangle.Fill>
                         </Rectangle>
                     </MenuItem.Icon>
